requirements.txt
django
psycopg2 
django-tailwind
django-allauth
django-widget-tweaks
jwt
PyJWT
cryptography
fido2
djangorestframework
Pillow
channels
requests
